{{cookiecutter.project_slug}}/backend/dependencies.py
import logging
from typing import Generator

from fastapi import Depends, HTTPException, status
from fastapi.security import OAuth2PasswordBearer
from jose import jwt
from pydantic import ValidationError
from sqlmodel import Session, select
import sys
import os

from {{cookiecutter.project_slug}}_core import schemas
from core import security
from core import config
from db.session import SessionLocal

reusable_oauth2 = OAuth2PasswordBearer(
    tokenUrl=f"{config.settings.API_V1_STR}/login/access-token"
)


def get_db() -> Generator:
    try:
        db = SessionLocal()
        yield db
    finally:
        db.close()


def get_current_user(
    db: Session = Depends(get_db), token: str = Depends(reusable_oauth2)
) -> schemas.User:
    try:
        print(f'TOKEN: {token}')
        payload = jwt.decode(
            token, config.settings.SECRET_KEY, algorithms=[security.ALGORITHM]
        )
        token_data = schemas.TokenPayload(**payload)
    except (jwt.JWTError, ValidationError):
        raise HTTPException(
            status_code=status.HTTP_403_FORBIDDEN,
            detail="Could not validate credentials",
        )
    user = select(schemas.User).where(
        schemas.User.id == token_data.sub)
    user = db.exec(user).first()
    print(f'CURRENT USER: {user.id}')
    if not user:
        raise HTTPException(status_code=404, detail="User not found")
    return user


def get_current_active_superuser(
    current_user: schemas.User = Depends(get_current_user),
) -> schemas.User:
    if not current_user.is_superuser:
        raise HTTPException(
            status_code=400, detail="The user doesn't have enough privileges"
        )
    return current_user


def get_settings(
    db: Session = Depends(get_db),
    current_user: schemas.User = Depends(get_current_user),
):
    settings = db.exec(select(schemas.Settings).where(schemas.Settings.user == current_user)).first()
    return settings
