{{cookiecutter.project_slug}}/backend/requirements.txt
aiofiles==0.6.0
aiohttp==3.7.3
alembic==1.4.3
async-timeout==3.0.1
asyncpg==0.21.0
attrs==20.3.0
bcrypt==3.2.0
bidict==0.21.2
broadcaster==0.2.0
cachetools==4.2.0
certifi==2020.12.5
cffi==1.14.4
chardet==3.0.4
click==7.1.2
cryptography==3.2.1
cssselect==1.1.0
cssutils==1.0.2
dnspython==2.0.0
ecdsa==0.14.1
email-validator==1.1.2
emails==0.6
exponent-server-sdk==1.0.0
fastapi
fastapi-contrib==0.2.9
h11==0.11.0
idna==2.10
jwt==1.1.0
lxml==4.6.2
Mako==1.1.3
MarkupSafe==1.1.1
motor==2.3.0
multidict==5.1.0
netifaces==0.10.6
numpy==1.19.5
passlib==1.7.4
premailer==3.7.0
psycopg2==2.8.6
pyasn1==0.4.8
pycparser==2.20
pydantic
pymongo==3.11.2
python-dateutil==2.8.1
python-dotenv==0.15.0
python-editor==1.0.4
python-engineio==4.0.0
python-jose==3.2.0
python-multipart==0.0.5
python-socketio==5.0.4
pytz==2020.5
requests==2.25.1
rsa==4.6
six==1.15.0
starlette==0.13.6
typing-extensions==3.7.4.3
ujson==4.0.1
urllib3==1.26.2
uvicorn==0.13.2
websocket-client==0.57.0
websockets==8.1
wincertstore==0.2
yarl==1.6.3
sqlmodel